backend/requirements.txt
ultralytics==8.3.38
Flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
Pillow==11.0.0
torch==2.5.1